+++ CHANGES.rst
@@ -5,7 +5,16 @@ threedi-qgis-plugin changelog
 0.9.4 (unreleased)
 ------------------
 
-- Nothing changed yet.
+- Add QGIS as dependency to Travis; make nosetests work on Travis for all
+  tests.
+
+- Add a new tool (``CacheClearer``) to clear the model cache.
+
+- Add an About tool class so that the about ``QAction`` can be added in the
+  same way as the other tools.
+
+- Add ``setup.cfg`` with coverage options (needs ``coverage``); exclude the
+  external and importer source files from tests.
 
 
 0.9.3 (2017-04-10)